+++ tests/test_layers/test_contrastive.py
@@ -2,8 +2,8 @@ import mobula.layers as L
 import numpy as np
 
 def test_contrastive():
-    A = np.array([[1,2,0], [2,1,0]]).astype(np.float)
-    B = np.array([[0,1,3], [2,3,1]]).astype(np.float)
+    A = np.array([[1,2,0], [2,1,0]]).astype(np.float32)
+    B = np.array([[0,1,3], [2,3,1]]).astype(np.float32)
 
     diff = A - B # 2 x 3
     dist_sq = np.sum(np.square(diff), 1).reshape((2,1)) # 2 x 1
